Delete dialog

+++ lib/chat/msg.dart
@@ -1,5 +1,6 @@
 import 'package:intl/intl.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:pocketbase/pocketbase.dart';
 
 import 'package:flutterchat/utils/pb_service.dart';
@@ -324,13 +325,42 @@ class MessageBubbleState extends State<MessageBubble> {
 
   // called when delete to message is selected
   Future<void> _handleDelete() async {
-    // TODO: ask confirmation (with dialog)
-
-    try {
-      await PocketBaseService().deleteMessage(widget.messageId);
-    } catch (e) {
-      if (mounted) snackBarText(context, 'Delete failed: ${e.toString()}');
-    }
+    showCupertinoDialog(
+      context: context,
+      builder: (BuildContext ctx) {
+        return CupertinoAlertDialog(
+          title: Text(localize("delete").toCapitalized()),
+          content: Text(localize("chat.delete.dialog")),
+          actions: [
+            // Yes button
+            CupertinoDialogAction(
+              onPressed: () {
+                setState(() async {
+                  try {
+                    await PocketBaseService().deleteMessage(widget.messageId);
+                  } catch (e) {
+                    if (mounted) snackBarText(context, 'Delete failed: ${e.toString()}');
+                  }
+                  Navigator.of(context).pop();
+                });
+              },
+              isDefaultAction: true,
+              isDestructiveAction: true,
+              child: Text(localize("yes").toCapitalized()),
+            ),
+            // No button
+            CupertinoDialogAction(
+              onPressed: () {
+                Navigator.of(context).pop();
+              },
+              isDefaultAction: false,
+              isDestructiveAction: false,
+              child: Text(localize("no").toCapitalized()),
+            )
+          ],
+        );
+      }
+    );
   }
 }
